+++ R/model-module-helpers.R
@@ -281,52 +281,6 @@ extract_sparse_matrix <- function(mat, prefix = "") {
   return(sparse_mat)
 }
 
-#' Simulate daily double censored PMF
-#'
-#' This function simulates the probability mass function of a  daily
-#' double-censored process. The process involves two distributions: a primary
-#' distribution which represents the censoring process for the primary event
-#' and another distribution (which is offset by the primary).
-#'
-#' @param max Maximum value for the computed CDF. If not specified, the maximum
-#' value is the maximum simulated delay.
-#' @param fun_primary Primary distribution function (default is \code{runif}).
-#' @param fun_dist Distribution function to be added to the primary (default is
-#' \code{rlnorm}).
-#' @param n Number of simulations (default is 1e6).
-#' @param primary_args List of additional arguments to be passed to the primary
-#' distribution function.
-#' @param dist_args List of additional arguments to be passed to the
-#' distribution function.
-#' @param ... Additional arguments to be passed to the distribution function.
-#' This is an alternative to `dist_args`.
-#'
-#' @return A numeric vector representing the PMF.
-#' @export
-#' @family modelmodulehelpers
-#' @examples
-#' simulate_double_censored_pmf(10, meanlog = 0, sdlog = 1)
-simulate_double_censored_pmf <- function(
-  max, fun_primary = stats::runif, primary_args = list(),
-  fun_dist = stats::rlnorm,
-  dist_args = list(...), n = 1e6, ...
-) {
-  lifecycle::deprecate_soft("0.5.0", "simulate_double_censored_pmf()")
-  cli::cli_inform(c(
-    "!" = "Users should instead use the {.pkg primarycensored} package for simulating double censored processes.", # nolint
-    "i" = "See {.url https://github.com/epinowcast/primarycensored} for more information." # nolint
-  ))
-  primary <- do.call(fun_primary, c(list(n), primary_args))
-  secondary <- primary + do.call(fun_dist, c(list(n), dist_args))
-  delay <- floor(secondary) - floor(primary)
-  if (missing(max)) {
-    max <- base::max(delay)
-  }
-  cdf <- ecdf(delay)(0:max)
-  pmf <- c(cdf[1], diff(cdf))
-  return(pmf)
-}
-
 #' Add maximum observed delay
 #'
 #' This function calculates and adds the maximum observed delay for each group